@@ -8,7 +8,7 @@ keywords = [
 ]
 license = {text = "Apache-2.0"}
 authors = [{name = "Edgar Riba", email = "edgar@kornia.org"}]
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 classifiers = [
   "Development Status :: 4 - Beta",
   "Environment :: Console",
@@ -21,7 +21,6 @@ classifiers = [
   "Natural Language :: English",
   "Operating System :: OS Independent",
   "Programming Language :: Python :: 3 :: Only",
-  "Programming Language :: Python :: 3.7",
   "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
@@ -80,52 +79,6 @@ namespaces = false
   skip-string-normalization = true
   skip-magic-trailing-comma = true
 
-[tool.pytest.ini_options]
-  addopts = "--color=yes"
-  testpaths = ["test"]
-  markers =[
-    "jit: mark a test as torchscript test",
-    "grad: mark a test as gradcheck test",
-    "nn: mark a test as module test"
-]
-
-[tool.coverage.report]
-exclude_lines = [
-  "pragma: no cover",
-  "def __repr__",
-  "if self.debug:",
-  "raise",
-  "if 0:",
-  "if __name__ == .__main__.:",
-]
-
-[tool.mypy]
-  files = ["kornia/"]
-  pretty = true
-  check_untyped_defs = true
-  disallow_any_generics = true
-  disallow_incomplete_defs = true
-  disallow_untyped_defs = true
-  show_error_codes = true
-  ignore_missing_imports = true
-  no_implicit_optional = true
-  warn_unused_ignores = true
-  warn_redundant_casts = true
-
-[tool.distutils.bdist_wheel]
-  universal = true
-
-[tool.pydocstyle]
-match = '.*\.py'
-ignore = """
-D105
-D107
-D203
-D204
-D213
-D406
-D407"""
-
 [tool.ruff]
 select = [
   "C9",
@@ -174,3 +127,49 @@ max-statements = 258
 "test/*" = ["PLR0402", "S101"]
 "test/enhance/test_core.py" = ["S311"]
 "test/nerf/test_camera_utils.py" = ["S310"]
+
+[tool.pytest.ini_options]
+  addopts = "--color=yes"
+  testpaths = ["test"]
+  markers =[
+    "jit: mark a test as torchscript test",
+    "grad: mark a test as gradcheck test",
+    "nn: mark a test as module test"
+]
+
+[tool.coverage.report]
+exclude_lines = [
+  "pragma: no cover",
+  "def __repr__",
+  "if self.debug:",
+  "raise",
+  "if 0:",
+  "if __name__ == .__main__.:",
+]
+
+[tool.mypy]
+  files = ["kornia/"]
+  pretty = true
+  check_untyped_defs = true
+  disallow_any_generics = true
+  disallow_incomplete_defs = true
+  disallow_untyped_defs = true
+  show_error_codes = true
+  ignore_missing_imports = true
+  no_implicit_optional = true
+  warn_unused_ignores = true
+  warn_redundant_casts = true
+
+[tool.distutils.bdist_wheel]
+  universal = true
+
+[tool.pydocstyle]
+match = '.*\.py'
+ignore = """
+D105
+D107
+D203
+D204
+D213
+D406
+D407"""
