@@ -239,7 +239,7 @@ def _get_precision(device: torch.device, dtype: Dtype) -> float:
 def _get_precision_by_name(
     device: torch.device, device_target: str, tol_val: float, tol_val_default: float = 1e-4
 ) -> float:
-    if device_target not in ['cpu', 'cuda', 'xla']:
+    if device_target not in ['cpu', 'cuda', 'xla', 'mps']:
         raise ValueError(f"Invalid device name: {device_target}.")
 
     if device_target in device.type:
